+++ R/nfunctions.R
@@ -41,13 +41,6 @@ ndoc.tokens <- function(x) {
     length(x)
 }
 
-#' @export
-#' @noRd
-ndoc.tokenizedTexts <- function(x) {
-    length(x)
-}
-
-
 #' @rdname ndoc
 #' @details \code{nfeature} returns the number of features from a dfm; it is an
 #'   alias for \code{ntype} when applied to dfm objects.  This function is only 
@@ -142,17 +135,17 @@ ntoken.character <- function(x, ...) {
 
 #' @noRd
 #' @export
+ntoken.tokens <- function(x, ...) {
+    lengths(x)
+}
+
+#' @noRd
+#' @export
 ntoken.dfm <- function(x, ...) {
     x <- as.dfm(x)
     if (length(list(...)) > 0)
         warning("additional arguments not used for ntoken.dfm()")
     rowSums(x)
-}
-
-#' @noRd
-#' @export
-ntoken.tokenizedTexts <- function(x, ...) {
-    lengths(x)
 }
 
 #' @noRd
@@ -172,20 +165,14 @@ ntype.corpus <- function(x, ...) {
 #' @export
 ntype.dfm <- function(x, ...) {
     x <- as.dfm(x)
-    ## only returns total NON-ZERO COUNT types
+    ## only returns total non-zero features
     rowSums(x > 0)
 }
 
 #' @noRd
 #' @export
-ntype.tokenizedTexts <- function(x, ...) {
-    vapply(lapply(x, unique), length, integer(1))
-}
-
-#' @export
-#' @noRd
-ntoken.tokens <- function(x, ...) {
-    lengths(x)
+ntype.tokens <- function(x, ...) {
+    sapply(unclass(x), function(y) length(unique(y[y > 0])))
 }
 
 #' count the number of sentences